+++ examples/simple.py
@@ -1,7 +1,9 @@
 import os
 import sys
+
 PATH = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.insert(0, PATH)
+
 from pyungo.core import Graph
 from pprint import pprint
 
@@ -9,15 +11,17 @@ from pprint import pprint
 if __name__ == '__main__':
     graph = Graph()
 
-    @graph.register(inputs=['a', 'b'], outputs=['c'])
+    @graph.register()
     def f_my_function(a, b):
-        return a + b
+        c = a + b
+        return c
 
     @graph.register(inputs=['d', 'a'], outputs=['e'])
     def f_my_function3(d, a):
-        return d - a
+        e = d - a
+        return e
 
-    @graph.register(inputs=['c'], outputs=['d'])
+    @graph.register(outputs=['d'])
     def f_my_function2(c):
         return c / 10.